+++ src/main/java/com/ururulab/ururu/order/controller/CartController.java
@@ -1,11 +1,11 @@
 package com.ururulab.ururu.order.controller;
 
 import com.ururulab.ururu.global.domain.dto.ApiResponseFormat;
-import com.ururulab.ururu.order.domain.dto.request.CartItemAddRequest;
-import com.ururulab.ururu.order.domain.dto.request.CartItemQuantityChangeRequest;
-import com.ururulab.ururu.order.domain.dto.response.CartItemAddResponse;
-import com.ururulab.ururu.order.domain.dto.response.CartResponse;
-import com.ururulab.ururu.order.domain.dto.response.CartItemQuantityChangeResponse;
+import com.ururulab.ururu.order.controller.dto.request.CartItemAddRequest;
+import com.ururulab.ururu.order.controller.dto.request.CartItemQuantityChangeRequest;
+import com.ururulab.ururu.order.controller.dto.response.CartItemAddResponse;
+import com.ururulab.ururu.order.controller.dto.response.CartResponse;
+import com.ururulab.ururu.order.controller.dto.response.CartItemQuantityChangeResponse;
 import com.ururulab.ururu.order.service.CartService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
