+++ src/main/java/com/ururulab/ururu/order/controller/OrderCreationController.java
@@ -1,9 +1,9 @@
 package com.ururulab.ururu.order.controller;
 
 import com.ururulab.ururu.global.domain.dto.ApiResponseFormat;
-import com.ururulab.ururu.order.domain.dto.request.CartOrderCreateRequest;
-import com.ururulab.ururu.order.domain.dto.request.GroupBuyOrderCreateRequest;
-import com.ururulab.ururu.order.domain.dto.response.OrderCreateResponse;
+import com.ururulab.ururu.order.controller.dto.request.CartOrderCreateRequest;
+import com.ururulab.ururu.order.controller.dto.request.GroupBuyOrderCreateRequest;
+import com.ururulab.ururu.order.controller.dto.response.OrderCreateResponse;
 import com.ururulab.ururu.order.service.OrderCreationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
