+++ src/main/java/com/ururulab/ururu/order/controller/dto/response/CartItemQuantityChangeResponse.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.order.domain.dto.response;
+package com.ururulab.ururu.order.controller.dto.response;
 
 /**
  * 장바구니 아이템 수량 변경 응답 DTO
