+++ src/main/java/com/ururulab/ururu/order/controller/dto/response/OrderCreateResponse.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.order.domain.dto.response;
+package com.ururulab.ururu.order.controller.dto.response;
 
 import java.util.List;
