+++ src/main/java/com/ururulab/ururu/order/service/CartService.java
@@ -4,12 +4,12 @@ import com.ururulab.ururu.groupBuy.domain.entity.GroupBuyOption;
 import com.ururulab.ururu.groupBuy.domain.repository.GroupBuyOptionRepository;
 import com.ururulab.ururu.member.domain.entity.Member;
 import com.ururulab.ururu.member.domain.repository.MemberRepository;
-import com.ururulab.ururu.order.domain.dto.request.CartItemAddRequest;
-import com.ururulab.ururu.order.domain.dto.request.CartItemQuantityChangeRequest;
-import com.ururulab.ururu.order.domain.dto.response.CartItemAddResponse;
-import com.ururulab.ururu.order.domain.dto.response.CartItemQuantityChangeResponse;
-import com.ururulab.ururu.order.domain.dto.response.CartItemResponse;
-import com.ururulab.ururu.order.domain.dto.response.CartResponse;
+import com.ururulab.ururu.order.controller.dto.request.CartItemAddRequest;
+import com.ururulab.ururu.order.controller.dto.request.CartItemQuantityChangeRequest;
+import com.ururulab.ururu.order.controller.dto.response.CartItemAddResponse;
+import com.ururulab.ururu.order.controller.dto.response.CartItemQuantityChangeResponse;
+import com.ururulab.ururu.order.controller.dto.response.CartItemResponse;
+import com.ururulab.ururu.order.controller.dto.response.CartResponse;
 import com.ururulab.ururu.order.domain.entity.Cart;
 import com.ururulab.ururu.order.domain.entity.CartItem;
 import com.ururulab.ururu.order.domain.repository.CartRepository;
