+++ src/main/java/com/ururulab/ururu/order/service/OrderCreationService.java
@@ -7,11 +7,11 @@ import com.ururulab.ururu.groupBuy.domain.entity.GroupBuyOption;
 import com.ururulab.ururu.groupBuy.domain.repository.GroupBuyOptionRepository;
 import com.ururulab.ururu.member.domain.entity.Member;
 import com.ururulab.ururu.member.domain.repository.MemberRepository;
-import com.ururulab.ururu.order.domain.dto.request.CartOrderCreateRequest;
-import com.ururulab.ururu.order.domain.dto.request.GroupBuyOrderCreateRequest;
-import com.ururulab.ururu.order.domain.dto.request.OrderItemRequest;
-import com.ururulab.ururu.order.domain.dto.response.OrderCreateResponse;
-import com.ururulab.ururu.order.domain.dto.response.OrderItemResponse;
+import com.ururulab.ururu.order.controller.dto.request.CartOrderCreateRequest;
+import com.ururulab.ururu.order.controller.dto.request.GroupBuyOrderCreateRequest;
+import com.ururulab.ururu.order.controller.dto.request.OrderItemRequest;
+import com.ururulab.ururu.order.controller.dto.response.OrderCreateResponse;
+import com.ururulab.ururu.order.controller.dto.response.OrderItemResponse;
 import com.ururulab.ururu.order.domain.entity.CartItem;
 import com.ururulab.ururu.order.domain.entity.Order;
 import com.ururulab.ururu.order.domain.entity.OrderItem;
@@ -25,8 +25,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Instant;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
